[cleanup] Decouple WFS capability storage class from GetCapabilities request class
