+++ src/providers/wfs/qgswfsdescribefeaturetype.cpp
@@ -15,7 +15,6 @@
 
 #include "qgswfsdescribefeaturetype.h"
 #include "moc_qgswfsdescribefeaturetype.cpp"
-#include "qgsmessagelog.h"
 #include <QUrlQuery>
 
 QgsWFSDescribeFeatureType::QgsWFSDescribeFeatureType( QgsWFSDataSourceURI &uri )
@@ -23,7 +22,7 @@ QgsWFSDescribeFeatureType::QgsWFSDescribeFeatureType( QgsWFSDataSourceURI &uri )
 {
 }
 
-bool QgsWFSDescribeFeatureType::requestFeatureType( const QString &WFSVersion, const QString &typeName, const QgsWfsCapabilities::Capabilities &caps )
+bool QgsWFSDescribeFeatureType::requestFeatureType( const QString &WFSVersion, const QString &typeName, const QgsWfsCapabilities &caps )
 {
   QUrl url( mUri.requestUrl( QStringLiteral( "DescribeFeatureType" ) ) );
   QUrlQuery query( url );
